+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 - **FIX**: Fix the resulting TextStyle applied to ShadTab
 - **FIX**: Fix the resulting decoration applied to ShadTab
 
+## 0.36.1
+
+- **FIX**: Expose `ShadButtonSizesTheme`.
+- **FEAT**: Add `textStyle` to `ShadButton` and `ShadButtonTheme` to customize the text style of the button.
+
 ## 0.36.0
 
 - **BREAKING CHANGE**: Remove `iconData` from `ShadAlert`, use `icon` instead.